Fix 64bit conversion warning.

+++ DLLDependencyViewer/DLLDependencyViewer/src/gui/search.cpp
@@ -95,7 +95,7 @@ void search(searcher& sch, string const* const& dll_name)
 	}
 
 	wchar_t* part;
-	DWORD len = SearchPathW(nullptr, tmpw.c_str(), nullptr, buff.size(), buff.data(), &part);
+	DWORD len = SearchPathW(nullptr, tmpw.c_str(), nullptr, static_cast<DWORD>(buff.size()), buff.data(), &part);
 	if(len != 0)
 	{
 		tmpw.assign(buff.data(), buff.data() + len);
